+++ customer.h
@@ -12,11 +12,11 @@ class Customer
     std::string secondName;
     short cardNum;
     short accountNum;
-    float balance;
+    double balance;
     Customers& owner;
     friend Customers;
 protected:
-    Customer(std::string firstName, std::string secondName, short cardNum, short accountNum, float balance, Customers& owner, Logger& logger);
+    Customer(std::string firstName, std::string secondName, short cardNum, short accountNum, double balance, Customers& owner, Logger& logger);
     void remove(Logger& logger);
 public:
     Customer(Customer& right);
@@ -26,14 +26,14 @@ public:
     void showCust();
     void setSecondName(std::string newSName, Logger& logger);
     void setFirstName(std::string newFName, Logger& logger);
-    void setBalance(float newBalance, Logger& logger);
+    void setBalance(double newBalance, Logger& logger);
     void setCardNum(std::string newCardNum, Logger& logger);
     void setAccountNum(std::string newAccountNum, Logger& logger);
 
 
     std::string getSecondName();
     std::string getFirstName();
-    float getBalance();
+    double getBalance();
     short getCardNum();
     short getAccountNum();
     int getID();
